requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.8
django-heroku==0.3.1
psycopg2==2.9.1
python-decouple==3.5
python-dotenv==0.17.0
pytz==2021.3
sqlparse==0.4.2
typing-extensions==3.10.0.2
whitenoise==5.3.0
